refactor(derive): Move into_app into coupled derives
